requirements.txt
netifaces==0.10.9
packaging==20.4
prompt-toolkit==3.0.5
pycryptodome==3.9.8
<<<<<<< HEAD
python-crontab==2.5.1
Pygments==2.6.1
PyNaCl==1.4.0
pytablewriter==0.54.0
python-dateutil==2.8.1
pytz==2020.1
=======
>>>>>>> upstream/master
requests==2.24.0
rich==9.10.0
python-rapidjson==0.9.1
ZODB==5.6.0
zodburi==2.4.0
jinja2==3.0.1
paramiko==2.7.2
